backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
pytest==8.2.0
httpx==0.27.0
pytest-asyncio==0.23.7
python-dotenv==1.0.1
alembic==1.13.1
pydantic-settings==2.2.1
pydantic[email]==2.7.1
anthropic==0.28.0
apscheduler==3.10.4
sentence-transformers>=2.2.0
pgvector>=0.2.0
